feat(test): format code
