+++ task/processor_test.go
@@ -1,6 +1,7 @@
 package task
 
 import (
+	pubsub "github.com/libp2p/go-libp2p-pubsub"
 	"github.com/machinefi/sprout/testutil"
 	"github.com/machinefi/sprout/types"
 	"github.com/machinefi/sprout/vm"
@@ -19,15 +20,15 @@ func TestNewProcessor(t *testing.T) {
 	ps := &p2p.PubSubs{}
 
 	t.Run("NewFailed", func(t *testing.T) {
-		patches = p2pNewPubSubs(patches, nil, errors.New(t.Name()))
+		patches = testutil.P2pNewPubSubs(patches, nil, errors.New(t.Name()))
 		_, err := NewProcessor(nil, nil, "", 0)
 		require.ErrorContains(err, t.Name())
 	})
-	patches = p2pNewPubSubs(patches, ps, nil)
+	patches = testutil.P2pNewPubSubs(patches, ps, nil)
 
 	t.Run("AddProjectFailed", func(t *testing.T) {
-		patches = projectManagerGetAllProjectID(patches, append([]uint64{}, 1))
-		patches = p2pPubSubsAdd(patches, errors.New(t.Name()))
+		patches = testutil.ProjectManagerGetAllProjectID(patches, append([]uint64{}, 1))
+		patches = testutil.P2pPubSubsAdd(patches, errors.New(t.Name()))
 		_, err := NewProcessor(nil, nil, "", 0)
 		require.ErrorContains(err, t.Name())
 	})
@@ -44,7 +45,7 @@ func TestReportFail(t *testing.T) {
 	patches = testutil.JsonMarshal(patches, []byte("any"), nil)
 
 	t.Run("PublishFailed", func(t *testing.T) {
-		patches = topicPublish(patches, errors.New(t.Name()))
+		patches = testutil.TopicPublish(patches, errors.New(t.Name()))
 		p.reportFail("taskID", errors.New(t.Name()), nil)
 	})
 }
@@ -60,7 +61,7 @@ func TestReportSuccess(t *testing.T) {
 	patches = testutil.JsonMarshal(patches, []byte("any"), nil)
 
 	t.Run("PublishFailed", func(t *testing.T) {
-		patches = topicPublish(patches, errors.New(t.Name()))
+		patches = testutil.TopicPublish(patches, errors.New(t.Name()))
 		p.reportSuccess("taskID", types.TaskStatePacked, "", nil)
 	})
 
@@ -96,29 +97,18 @@ func TestProcessorHandleP2PData(t *testing.T) {
 			TaskStateLog: nil,
 		}
 		patches = processorReportSuccess(patches)
-		patches = projectManagerGet(patches, errors.New(t.Name()))
+		patches = testutil.ProjectManagerGet(patches, errors.New(t.Name()))
 		patches = processorReportFail(patches)
 		p.handleP2PData(data, nil)
 	})
-	patches = projectManagerGet(patches, nil)
+}
 
-	//t.Run("ProofFailed", func(t *testing.T) {
-	//	data := &p2p.Data{
-	//		Task: &types.Task{
-	//			ID: "",
-	//			Messages: []*types.Message{{
-	//				ID:             "id1",
-	//				ProjectID:      uint64(0x1),
-	//				ProjectVersion: "0.1",
-	//				Data:           "data",
-	//			}},
-	//		},
-	//		TaskStateLog: nil,
-	//	}
-	//	//patches = processorReportSuccess(patches)
-	//	//patches = processorReportFail(patches)
-	//	patches = vmHandlerHandle(patches, errors.New(t.Name()))
-	//	p.handleP2PData(data, nil)
-	//})
+func processorReportSuccess(p *Patches) *Patches {
+	var pro *Processor
+	return ApplyPrivateMethod(pro, "reportSuccess", func(taskID string, state types.TaskState, comment string, topic *pubsub.Topic) {})
+}
 
+func processorReportFail(p *Patches) *Patches {
+	var pro *Processor
+	return ApplyPrivateMethod(pro, "reportFail", func(taskID string, err error, topic *pubsub.Topic) {})
 }
