+++ testutil/patches_mod_persistence.go
@@ -1,0 +1,34 @@
+package testutil
+
+import (
+	"fmt"
+	"reflect"
+	"time"
+
+	. "github.com/agiledragon/gomonkey/v2"
+	"github.com/machinefi/sprout/persistence"
+	"github.com/machinefi/sprout/types"
+)
+
+func PersistencePostgresUpdateState(p *Patches, err error) *Patches {
+	var pg *persistence.Postgres
+	return p.ApplyMethodFunc(
+		reflect.TypeOf(pg),
+		"UpdateState",
+		func(taskID string, state types.TaskState, comment string, createdAt time.Time) error {
+			fmt.Println("up")
+			return err
+		},
+	)
+}
+
+func PersistencePostgresFetchByID(p *Patches, task *types.Task, err error) *Patches {
+	var pg *persistence.Postgres
+	return p.ApplyMethodFunc(
+		reflect.TypeOf(pg),
+		"FetchByID",
+		func(taskID string) (*types.Task, error) {
+			return task, err
+		},
+	)
+}
